// frontend/src/types/api/listAppliedLabelsApiLabelsGetSort.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * mailassist
 * OpenAPI spec version: 0.1.0
 */

export type ListAppliedLabelsApiLabelsGetSort = typeof ListAppliedLabelsApiLabelsGetSort[keyof typeof ListAppliedLabelsApiLabelsGetSort];


export const ListAppliedLabelsApiLabelsGetSort = {
  newest: 'newest',
  oldest: 'oldest',
  label: 'label',
} as const;